chore: refactor certificate.
